puzzle_complete!

+++ str/apps/include/BoardAnalyzer.h
@@ -9,17 +9,17 @@ class BoardAnalyzer
 public:
 	BoardAnalyzer();
 	~BoardAnalyzer();
-	void set(int8_t *initPositions);
-	int8_t getInLeftSquareCount();
+	void set(int *initPositions);
+	int getInLeftSquareCount();
 	bool hasExchangePatternOnLeftSquare();
 	bool hasExchangePatternOnPentagon();
-	int8_t getInRightSideCount();
+	int getInRightSideCount();
 	
 private:
-	int8_t *positions;
-	const int8_t leftSquare[5] = {0, 1, 5, 2, 10};
-	const int8_t pentagon[5] = {8, 14, 13, 9, 6};
-	const int8_t rightSide[6] = {3, 4, 7, 9, 11, 15};
+	int *positions;
+	const int leftSquare[5] = {0, 1, 5, 2, 10};
+	const int pentagon[5] = {8, 14, 13, 9, 6};
+	const int rightSide[6] = {3, 4, 7, 9, 11, 15};
 	const BlockColor BlockArea[16] = {
 		BlockColor::Black,
 		BlockColor::Red,
